+++ CHANGES.rst
@@ -10,6 +10,7 @@ in progress
     - Add translations into Norwegian, Catalan and Spanish
     - Add translations into Portuguese and Swedish
     - Don't lower case language name for honoring labels like "nb_NO"
+    - Improve language/variant chooser page significantly
 
 2018-02-08 0.4.0
 ----------------